server/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator
jinja2
httpx
python-jose
python-multipart
